cleanup, change nano id to funny names
