+++ models.go
@@ -1,0 +1,23 @@
+package main
+
+import (
+	"github.com/gorilla/websocket"
+)
+
+type Client struct {
+    conn *websocket.Conn
+}
+
+type ForwardedRequest struct {
+    Method  string              `json:"method"`
+    URL     string              `json:"url"`
+    Headers map[string][]string `json:"headers"`
+    Body    string              `json:"body"`
+}
+
+type ForwardedResponse struct {
+    StatusCode int               `json:"status_code"`
+    Headers    map[string][]string `json:"headers"`
+    Body       string            `json:"body"`
+}
+
